DNASeqFISHChromosomeAssignment.jl

This package implements the Longest Disjoint Paths algorithm for identifying chromosome strands in DNA SeqFISH+ data. The intuition behind this method DNA SeqFISH+ is like imaging beads on an invisible string, where the loci we measure are the beads and the genomic DNA is the string. We know the spatial and genomic coordinates of each locus we find, so we can infer that DNA strands connect loci that are close both spatially and genomically.

However, we do not try to model the branching nature of replicating chromosomes. To correctly classify replicated loci on chromosome pairs that are spatially distinct, we combine the LDP chromosome identification method with DBSCAN.

DNASeqFISHChromosomeAssignment.assign_chromosomes — Function

assign_chromosomes(pnts :: DataFrame, params :: ChromSepParams, optimizer)

This function assigns genomic loci from DNA seqFISH+ data to indiviual chromosomes using DBSCAN and Longest Disjoint Paths (LDP). First, genomic loci are clustering using DBSCAN. Then DBSCAN clusters that contain a proportion of unique genomic loci less than the minpropunique parameter set in the params object and whose average spatial distance between subsequent spatial loci is greater than the r_ldp parameter (also et in the params object) will be split, using the LDP algorithm. If these conditions are not met, then we find one LDP that is a subset of the cluster.

LDP works by finding the longest one or two weighted paths of loci of increasing genomic coordinate within the a DBSCAN cluster. More specifically, it builds a directed acyclic graph where every locus has an out going edge to genomically down stream loci within the search radius. Edges are weighted by:

$w = \frac{1}{g_p - g_c} × e^{-\frac{||x_p - x_c||^2}{2σ^2}}$

where x denotes a vector of spatial coordinates, g denotes genomic coordinates, and p and c subscripts denote parent and child nodes respectively. LDP then uses Integer Programming to find the one or two disjoint paths that maximize the sum of the weights of all edges in the paths.

Arguments

df: Takes a DataFrame (pnts) of decoded DNA seqFISH data where each row describes a locus. Must include columns:

- fov : the field of view in which locus was imaged
- cellID : the cell in which the locus was imaged
- chrom : chromosome on which locus is located
- x : x spatial coordinate of locus
- y : y spatial coordinate of locus
- z : z spatial coordinate of locus
- g : genomic coordinate of locus

prms : ChromSepParams object containing parameters.

optimizer : optional - JuMP compatible optimizer for performing Longest Disjoint Paths optimization. Uses GLPK by default.

returns a copy of the input points sorted by fov, cellID, chrom, and g with a new columns: dbscan_allele : DBSCAN clusters of each locus dbscan_ldp_allele : ldp subclusters of each dbscan cluster dbscan_ldp_nbr_allele : dbscan clusters that are not split, and expands split ldp clusters by assigning ldp neighbors to their cluster.

allele = -1 means that a locus was not assigned to a chromosome. allele > 0 indicates the allele number of the chromosome that the locus was assigned to.

DNASeqFISHChromosomeAssignment.ChromSepParams — Type

mutable struct ChromSepParams(
    r_dbscan :: Real
    r_ldp :: Real
    r_ldp_nbr :: Real
    sigma :: Real
    min_size :: Int64
    dbscan_min_nbrs :: Int64
    min_prop_unique :: Real
    )

r_dbscan : the DBSCAN search radius

r_ldp : the longest disjoint path search radius

r_ldp_nbr : the search radius to add neighboring loci to LDP clusters after separating a DBSCAN cluster

sigma : defines the weight function for how much to penalize grouping adjacent genomic loci onto the the same chromosome based on their spatial distance from each other.

min_size : the minimum acceptable number of loci on a strand.

dbscan_min_pnts : the minimum allowed number of points for a DBSCAN cluster. DBSCAN clusters with fewer points are discarded.

min_prop_unique : the minimum proportion of unique loci a DBSCAN cluster is allowed to have before considering splitting it into longest disjoint paths.

DNASeqFISHChromosomeAssignment.ChromSepParams — Method

ChromSepParams()

Constructor for ChromSepParams object using the default parameter values. The user can set the parameter values using parameter setting methods.

DNASeqFISHChromosomeAssignment.set_r_dbscan — Function

set_r_dbscan(p :: ChromSepParams, r :: Real)

Set the DBSCAN search radius for clustering loci.

DNASeqFISHChromosomeAssignment.set_r_ldp — Function

set_r_ldp(p :: ChromSepParams, r :: Real)

Set the longest disjoint path search radius for clustering loci.

DNASeqFISHChromosomeAssignment.set_r_ldp_nbr — Function

set_r_ldp_nbr(p :: ChromSepParams, r :: Real)

Set the search radius for clustering loci near LDPs after separating DBSCAN clusters with LDPs.

DNASeqFISHChromosomeAssignment.set_sigma — Function

set_sigma(p :: ChromSepParams, s :: Real)

Set the sigma for the spatial distance penalty in the longest disjoint paths optimization.

DNASeqFISHChromosomeAssignment.set_min_size — Function

set_min_size(p :: ChromSepParams, min_size :: Real)

Set the minimum size of DBSCAN clusters and minimum length of the longest disjoint paths.

DNASeqFISHChromosomeAssignment.set_dbscan_min_nbrs — Function

set_dbscan_min_nbrs(p :: ChromSepParams, mns :: Int64)

Set the minimum number of neighbors a locus must have within the DBSCAN radius to be incuded in a DBSCAN cluster.

DNASeqFISHChromosomeAssignment.set_minProp_unique — Function

set_minProp_unique(p :: ChromSepParams, mpu :: Real)

If the proportion of unique loci in a DBSCAN cluster is less than this, the algorithm will then check whether the average spatial distance between loci of subsequent genomic coordinates is greater than r_ldp. If so, then the DBSCAN cluster is split into to Longest Disjoint paths.